Add component to fetch a registered model (#27)
Users may well have a pretrained and registered model which they want to use in a model analysis. This adds a 'Fetch Registered Model' component to help with this scenario. This takes the id of the desired model, and writes out the JSON file which the RAI Analysis 'constructor' component required.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>