Clientside SDK (#28)
Since we're assembling things across multiple Runs, need a small clientside SDK. For now, this needs to depend on some V1 SDK bits as well, unfortunately. In addition, extra permissions are needed on the storage account, so that the data plane can be accessed. If the user created the storage account (even indirectly), then they will have permissions to give themselves this permission, but it's not added by default.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>