Verify insights came from same constructor (#30)
Add a check to ensure that all the insights being assembled came from the same constructor. This includes a test which runs a pipeline which is expected to fail. Unfortunately, there is not currently a good way of checking _why_ the pipeline failed.

Also, rerun `black` on the affected files.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>